backend/requirements.txt
google-cloud-tasks
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic-settings
google-cloud-firestore
google-cloud-storage